requirements.txt
streamlit
google-generativeai
groq
langchain
langchain-groq
python-dotenv